Add Jekyll Site
I created a Jekyll site under `site`. It is built using Zurb Foundation
and broken into a few includes and layouts.

Foundation style overrides take place in `site/css/site.css`, but trying
to keep things as simple as possible for now. Right now it is just a few
color style overrides. I went with a green and white theme because I
thought it would be a fun play on the red-green-refactor thing. And red
is a bit overplayed in ruby sites.

I am using pygments for highlighting. Unfortunately Jekyll doesn't seem
to support this for fenced code blocks in markdown. Instead we have to
use the ugly:

    {% highlight ruby %}
      puts 'foo'
    {% endhighlight %}

Ugh. But it was fewer moving parts this way, and I had a drop in syntax
highlighting theme.

I also created a `CNAME` that points to your GitHub domain. I believe
you will need to do the following:

1. Create a `gh-pages` branch
1. Make `gh-pages` the default branch
1. Re-push like a boss

DNS is setup per [GitHub
Help](https://help.github.com/articles/setting-up-a-custom-domain-with-pages).